requirements.txt
auto_gptq==0.2.2
beautifulsoup4==4.12.2
chromadb==0.3.23
click==8.1.6
Flask==2.3.2
huggingface_hub==0.16.4
langchain==0.0.263
llama_cpp_python==0.1.77
matplotlib==3.7.2
numpy==1.23.5
numpy==1.25.2
psycopg2==2.9.6
psycopg2_binary==2.9.6
python-dotenv==1.0.0
scikit_learn==1.3.0
selenium==4.11.2
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
